agent/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bugagent"
version = "0.1.0"
description = "AI agent that monitors Telegram chats and creates Taiga bug tasks"
requires-python = ">=3.12"
authors = [{ name = "BugAgent" }]

dependencies = [
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.67",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.2.0",
    # MCP client
    "mcp>=1.0.0",
    # Qdrant
    "qdrant-client[fastembed]>=1.11.0",
    "fastembed>=0.4.0",
    # HTTP / async
    "httpx>=0.27.0",
    "aiohttp>=3.10.0",
    "aiofiles>=24.1.0",
    # RabbitMQ
    "aio-pika>=9.4.0",
    # Config & utilities
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "structlog>=24.4.0",
    # Retries and resilience
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
